refactor: move managed service auth data structs to config module
